pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["pptx704 <rafeedm.bhuiyan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {version = "^0.110.0", extras = ["uvicorn"]}
httpx = "^0.27.0"
serpapi = "^0.1.5"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
xmltodict = "^0.13.0"
bs4 = "^0.0.2"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
